php73  => php72, disabled redis, mongodb

+++ scripts/sync_config_files.sh
@@ -1,27 +1,53 @@
 #!/usr/bin/env bash
 
-CURRENT_DIR=$(dirname "$0")
+# CURRENT_DIR=$(dirname "$0")
+CURRENT_DIR="/home/vagrant/scripts"
 
 # Apache config
-cp $CURRENT_DIR/../conf/apache2.conf /etc/apache2/
+if [ -d "/etc/apache2/" ]
+then
+    cp $CURRENT_DIR/../conf/apache2.conf /etc/apache2/
+else
+    echo "Apache config file not found."
+fi
 
 # Mysql config
-cp $CURRENT_DIR/../conf/mysqld.cnf /etc/mysql/mysql.conf.d/
+if [ -d "/etc/mysql/mysql.conf.d/" ]
+then
+    cp $CURRENT_DIR/../conf/mysqld.cnf /etc/mysql/mysql.conf.d/
+else
+    echo "Mysql config folder not exists."
+fi
 
 # Redis config
-# cp $CURRENT_DIR/../conf/redis.conf /etc/redis/
+if [ -d "/etc/redis/" ]
+then
+    cp $CURRENT_DIR/../conf/redis.conf /etc/redis/
+else
+    echo "Redis config folder not exists."
+fi
 
 # Memcached config
 cp $CURRENT_DIR/../conf/memcached.conf /etc/
 
 # Apache vhost config
-rm -rf /etc/apache2/sites-enabled/*
-cp -r $CURRENT_DIR/../sites-enabled/ /etc/apache2/
+if [ -d /etc/apache2/sites-enabled/ ]
+then
+    rm -rf /etc/apache2/sites-enabled/*
+	cp -r $CURRENT_DIR/../sites-enabled/ /etc/apache2/
+else
+    echo "Apache sites-enabled folder not exists."
+fi
 
 # Apache config
-cp $CURRENT_DIR/../conf/php56.conf /etc/apache2/
+if [ -d /etc/apache2/ ]
+then
+    cp $CURRENT_DIR/../conf/php56.conf /etc/apache2/
+else
+    echo "Apache folder not exists."
+fi
 
 systemctl restart apache2
 systemctl restart mysql
 /etc/init.d/memcached restart
-systemctl restart redis
+# systemctl restart redis
